+++ src/main/java/com/inditex/ecommerce/domain/repositories/PriceRepositoryPort.java
@@ -6,5 +6,5 @@ import java.time.LocalDate;
 import java.util.List;
 
 public interface PriceRepositoryPort {
-    public List<Price> find(Long brandId, Long productId, LocalDate applicationDate);
+    public List<Price> find(Integer brandId, Integer productId, LocalDate applicationDate);
 }
